Revert "Altimeter: fix output format for hPa/InHg settings"
Replaced by some line of Nasal, that way we don't touch the internal precision of the altimeter

This reverts commit 84a26de9ae9ade8629dee53cac8b9357f0930690.

+++ src/Instrumentation/altimeter.hxx
@@ -43,8 +43,8 @@ public:
 
     double getSettingInHg() const;
     void setSettingInHg( double value );
-    int getSettingHPa() const;
-    void setSettingHPa( int value );
+    double getSettingHPa() const;
+    void setSettingHPa( double value );
 
 private:
     std::string _name;
